Create INPUT-Coin
The provided Solidity smart contract is named "INT" and is an implementation of an ERC-20 token. It inherits functionalities from the "ERC20.sol" contract in the OpenZeppelin library, which is a popular framework for building Ethereum smart contracts.

The contract sets the license identifier to MIT, which means it is distributed under the permissive MIT License, allowing users to utilize, modify, and distribute the code as long as they include the original copyright and disclaimer notice.

Upon deployment, the contract's constructor function is called, creating a new ERC-20 token with the name "INT Token" and symbol "INT." The contract creator (deployer) is assigned 1,000,000 INT tokens as the initial supply.